requirements.txt
flask
flask-cors
google-generativeai
PyMuPDF
gunicorn
requests